+++ src/Routing/RedirectableRequestMatcher.php
@@ -12,16 +12,28 @@
 namespace Symfony\Cmf\Bundle\RoutingBundle\Routing;
 
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Routing\Exception\ExceptionInterface;
 use Symfony\Component\Routing\Exception\ResourceNotFoundException;
 use Symfony\Component\Routing\Matcher\RequestMatcherInterface;
+use Symfony\Component\Routing\RequestContext;
 
+/**
+ * Reproduce the behaviour of the Symfony router to redirect paths with a trailing slash if necessary.
+ *
+ * The logic is taken from Symfony\Component\Routing\Matcher\RedirectableUrlMatcher and Symfony\Bundle\FrameworkBundle\Routing\RedirectableCompiledUrlMatcher
+ *
+ * @see https://symfony.com/doc/4.1/routing.html#routing-trailing-slash-redirection
+ */
 class RedirectableRequestMatcher implements RequestMatcherInterface
 {
     private RequestMatcherInterface $decorated;
 
-    public function __construct(RequestMatcherInterface $decorated)
+    private RequestContext $requestContext;
+
+    public function __construct(RequestMatcherInterface $decorated, RequestContext $requestContext)
     {
         $this->decorated = $decorated;
+        $this->requestContext = $requestContext;
     }
 
     /**
@@ -45,18 +57,25 @@ class RedirectableRequestMatcher implements RequestMatcherInterface
 
             $pathinfo = $trimmedPathinfo === $pathinfo ? $pathinfo.'/' : $trimmedPathinfo;
 
-            $parameters = $this->decorated->matchRequest($this->rebuildRequest($request, $pathinfo));
+            try {
+                $parameters = $this->decorated->matchRequest($this->rebuildRequest($request, $pathinfo));
+            } catch (ExceptionInterface $e2) {
+                throw $e;
+            }
 
-            return $this->redirect($pathinfo, $parameters['_route'] ?? null) + $parameters;
+            return $this->redirect($pathinfo, $parameters['_route'] ?? null, $this->requestContext->getScheme()) + $parameters;
         }
     }
 
-    private function redirect(string $path, string $route): array
+    private function redirect(string $path, string $route, string $scheme = null): array
     {
         return [
             '_controller' => 'Symfony\\Bundle\\FrameworkBundle\\Controller\\RedirectController::urlRedirectAction',
             'path' => $path,
             'permanent' => true,
+            'scheme' => $scheme,
+            'httpPort' => $this->requestContext->getHttpPort(),
+            'httpsPort' => $this->requestContext->getHttpsPort(),
             '_route' => $route,
         ];
     }